Clarify wording
I am a bit overly pedantic here, but ECEF is just a Cartesian system that rotates together with the planet. One may define a datum, or one may not, and even then, WGS84 is just one datum for Earth, while ECEF exists also for Mars, etc., and even bodies that are not approximately spherical.